abschluss und readme

+++ ft_strmapi.c
@@ -6,15 +6,15 @@
 /*   By: eboualla <eboualla@student.42vienna.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/23 14:28:40 by eboualla          #+#    #+#             */
-/*   Updated: 2026/04/23 14:45:33 by eboualla         ###   ########.fr       */
+/*   Updated: 2026/05/01 15:20:53 by eboualla         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 char	*ft_strmapi(const char *s, char (*f)(unsigned int, char))
 {
-	size_t	i;
-	char	*arr;
+	unsigned int	i;
+	char			*arr;
 
 	i = 0;
 	if (!s || !f)
